refactor: narrow StreamId from u64 to u32 to reduce wire overhead
StreamId(u64) was oversized — u32 supports 4B streams which is more
than sufficient. This saves 4 bytes per frame on every opcode that
carries stream_id (nearly all of them), reducing network overhead.

- Change StreamId inner type from u64 to u32
- Update all wire encode/decode (put_u64→put_u32, get_u64→get_u32)
- Update all variable header size calculations
- Change MySQL schema from BIGINT to INT UNSIGNED (in-place, dev phase)
- Change metadata binds from `stream_id.0 as i64` to `stream_id.0`
  (properly typed u32 for INT UNSIGNED, no lossy casts)
- Change metadata reads from get::<i64> to get::<u32>
- Update AtomicU64→AtomicU32 for next_stream_id counter
- S3 extent header keeps stream_id as u64 (on-disk format stability)
- Update all tests, benchmarks, integration tests, design.md

+++ components/common/src/types.rs
@@ -48,7 +48,7 @@ pub const FLAG_SYSTEM_TICK: u8 = 0x02;
 
 /// Unique identifier for a stream.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct StreamId(pub u64);
+pub struct StreamId(pub u32);
 
 impl Display for StreamId {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
